Move blur fade delay config
