+++ src/components/ui/BlurFade.tsx
@@ -64,6 +64,4 @@ const BlurFade = ({
   );
 };
 
-export const BLUR_FADE_DELAY = 0.04;
-
 export default BlurFade;
